refactor: move currency conversion to utils
