+++ presets/forex_rates.json
@@ -9,5 +9,5 @@
   "HKD": {"USD": 0.128},
   "AUD": {"USD": 0.65},
   "THB": {"USD": 0.028},
-  "IDR": {"USD": 0.000062},
+  "IDR": {"USD": 0.000062}
 }
